Move all new functional linked with stp/OOB positions from parallel_backend_sycl_reduce_then_scan.h into parallel_backend_sycl_reduce_then_scan_pos_tools.h
