+++ include/oneapi/dpl/pstl/hetero/dpcpp/parallel_backend_sycl_reduce_then_scan.h
@@ -36,6 +36,7 @@
 #include "../../tuple_impl.h"
 #include "../../utils_ranges.h"
 #include "../utils_hetero.h"
+#include "parallel_backend_sycl_reduce_then_scan_pos_tools.h"
 
 namespace oneapi
 {
@@ -75,21 +76,6 @@ struct __temp_data_array
     oneapi::dpl::__internal::__lazy_ctor_storage<_ValueT> __data[elements];
 };
 
-template <typename, typename = void>
-struct __select_max_outputs_per_input : std::integral_constant<std::uint16_t, 1>
-{
-    // By default, each work-item emits at most one output per scanned element.
-};
-
-template <typename _T>
-struct __select_max_outputs_per_input<_T, std::void_t<decltype(_T::__max_outputs_per_input)>>
-    : std::integral_constant<std::uint16_t, _T::__max_outputs_per_input>
-{
-};
-
-template <typename _T>
-constexpr std::uint16_t __select_max_outputs_per_input_v = __select_max_outputs_per_input<std::decay_t<_T>>::value;
-
 // This is a stand-in for a temporary data structure which is used to turn set() into a no-op. This is used in the case
 // where no temporary register data is needed within reduce then scan kern
 struct __noop_temp_data
@@ -100,47 +86,6 @@ struct __noop_temp_data
     {
     }
 };
-
-// Temporary data stand-in which discards the stored values and instead captures
-// the source position of the element at a specific index during a reduce then scan operation.
-template <typename _SrcDataPosT>
-struct __src_pos_capturing_temp_data
-{
-  public:
-    __src_pos_capturing_temp_data(std::uint16_t __idx_for_src_pos) : __idx_for_src_pos(__idx_for_src_pos) {}
-
-    template <typename _ValueT2>
-    void
-    set(std::uint16_t __idx, const _ValueT2&, _SrcDataPosT __src_idx)
-    {
-        if (__idx == __idx_for_src_pos)
-            __saved_src_pos = __src_idx;
-    }
-
-    _SrcDataPosT
-    __get_saved_src_pos() const
-    {
-        return __saved_src_pos;
-    }
-
-  private:
-    const std::uint16_t __idx_for_src_pos = 0;
-    _SrcDataPosT __saved_src_pos = {};
-};
-
-template <typename = void>
-struct __temp_data_capture_indexes_flag : std::false_type
-{
-};
-
-template <typename _SrcDataPosT>
-struct __temp_data_capture_indexes_flag<__src_pos_capturing_temp_data<_SrcDataPosT>> : std::true_type
-{
-};
-
-template <typename _TempData>
-inline constexpr bool __temp_data_capture_indexes_flag_v =
-    __temp_data_capture_indexes_flag<std::decay_t<_TempData>>::value;
 
 // Extracts a range from a zip iterator based on the element ID
 template <std::size_t _EleId>
@@ -576,7 +521,7 @@ __set_generic_operation_iteration(const _InRng1& __in_rng1, const _InRng2& __in_
     using _ValueTypeRng2 = typename oneapi::dpl::__internal::__value_t<_InRng2>;
 
     auto __write_temp_element = [&](const _SizeType __count_arg, const auto& __value) {
-        if constexpr (__temp_data_capture_indexes_flag_v<_TempOutput>)
+        if constexpr (__internal::__temp_data_capture_indexes_flag_v<_TempOutput>)
             __temp_out.set(__count_arg, __value, {__idx1, __idx2});
         else
             __temp_out.set(__count_arg, __value);
@@ -653,15 +598,6 @@ __set_generic_operation_iteration(const _InRng1& __in_rng1, const _InRng2& __in_
     }
 }
 
-// Tag type to indicate that no callback is provided
-struct __no_callback_tag
-{
-};
-
-// Helper variable template to check if a callback is provided or not
-template <typename _TCallback>
-inline constexpr bool __is_no_callback_v = std::is_same_v<std::decay_t<_TCallback>, __no_callback_tag>;
-
 // Set operation generic implementation, used for serial set operation of intersection, difference, union, and
 // symmetric difference.
 template <bool _CopyMatch, bool _CopyDiffSetA, bool _CopyDiffSetB>
@@ -700,7 +636,7 @@ struct __set_generic_operation
                const _SizeType __num_eles_min, _TempOutput& __temp_out, const _Compare __comp, _Proj1 __proj1,
                _Proj2 __proj2, _FinalPosSaver __final_pos_saver) const
     {
-        if constexpr (__is_no_callback_v<_FinalPosSaver>)
+        if constexpr (__internal::__is_no_callback_v<_FinalPosSaver>)
         {
             return __check_bounds_and_run_loop(__in_rng1, __in_rng2, __idx1, __idx2, __num_eles_min, __temp_out, __comp,
                                                __proj1, __proj2);
@@ -1094,20 +1030,12 @@ struct __gen_set_op_from_known_balanced_path
     _Proj2 __proj2;
 };
 
-template <typename, typename = void>
-struct __detect_oob_in_two_steps_selector : std::false_type
-{
-};
-
 template <typename _SetOpCount, typename _TempData, typename _RetType, typename _Compare, typename _Proj1,
           typename _Proj2>
-struct __detect_oob_in_two_steps_selector<
+struct __internal::__detect_oob_in_two_steps_selector<
     __gen_set_op_from_known_balanced_path<_SetOpCount, _TempData, _RetType, _Compare, _Proj1, _Proj2>> : std::true_type
 {
 };
-
-template <typename _T>
-inline constexpr bool __detect_oob_in_two_steps_v = __detect_oob_in_two_steps_selector<std::decay_t<_T>>::value;
 
 // kernel for balanced path to partition the input into tiles by calculating balanced path on diagonals of tile bounds
 template <typename _GenInput, typename _KernelName>
@@ -1697,8 +1625,8 @@ struct __temp_data_required<_T, std::void_t<typename _T::TempData>>
 
 template <bool _Bounded, bool __is_inclusive, bool __is_unique_pattern_v, typename _GenInput,
           typename _ScanInputTransform, typename _BinaryOp, typename _WriteOp, typename _ValueType, typename _InRng,
-          typename _OutRng, typename _CommTag, typename _OnOOBReached = __no_callback_tag,
-          typename _FinalPosSaver = __no_callback_tag>
+          typename _OutRng, typename _CommTag, typename _OnOOBReached = __internal::__no_callback_tag,
+          typename _FinalPosSaver = __internal::__no_callback_tag>
 void
 __scan_through_elements_helper(const sycl::nd_item<1>& __ndi, _GenInput __gen_input,
                                _ScanInputTransform __scan_input_transform, _BinaryOp __binary_op, _WriteOp __write_op,
@@ -1748,7 +1676,7 @@ __scan_through_elements_helper(const sycl::nd_item<1>& __ndi, _GenInput __gen_in
                 // this many writes per scanned element, otherwise the unchecked write path could be selected for
                 // set operations and overrun __out_rng (corrupting memory and skipping OOB position detection).
                 const std::size_t __max_writes_this_sub_group =
-                    std::size_t{__iters_per_item} * __sub_group_size * __select_max_outputs_per_input_v<_TempData>;
+                    std::size_t{__iters_per_item} * __sub_group_size * __internal::__select_max_outputs_per_input_v<_TempData>;
                 if (__carry_in + __max_writes_this_sub_group + __is_unique_pattern_v > __out_rng_size)
                 {
                     auto __bounded_write_op = [&](std::size_t __id, const auto& __v) {
@@ -1830,29 +1758,6 @@ class __reduce_then_scan_reduce_kernel;
 template <typename... _Name>
 class __reduce_then_scan_scan_kernel;
 
-// Sentinel type used as a stand-in for the stop-position accessor when _Bounded=false.
-struct __no_stop_pos_acc_tag
-{
-};
-
-template <typename _T>
-inline constexpr bool __is_no_stop_pos_acc_v = std::is_same_v<std::remove_cv_t<_T>, __no_stop_pos_acc_tag>;
-
-template <bool _Bounded, typename _StopPosStorage>
-auto
-__get_stop_pos_accessor_opt(sycl::handler& __cgh, _StopPosStorage& __stop_pos_storage)
-{
-    if constexpr (_Bounded)
-    {
-        // By using this sycl::read_write option we implement source data initialization under this accessor
-        return __get_accessor(sycl::read_write, __stop_pos_storage, __cgh, __dpl_sycl::__no_init{});
-    }
-    else
-    {
-        return __no_stop_pos_acc_tag{};
-    }
-}
-
 template <bool _Bounded, bool __is_inclusive, bool __is_unique_pattern_v, typename _ScanOpsTag,
           typename _GenReduceInput, typename _ReduceOp, typename _InitType, typename _KernelName>
 struct __parallel_reduce_then_scan_reduce_submitter;
@@ -1883,7 +1788,7 @@ struct __parallel_reduce_then_scan_reduce_submitter<_Bounded, __is_inclusive, __
             __cgh.depends_on(__prior_event);
             oneapi::dpl::__ranges::__require_access(__cgh, __in_rng);
             auto __temp_acc = __get_accessor(sycl::write_only, __scratch_container, __cgh, __dpl_sycl::__no_init{});
-            auto __stop_pos_acc = __get_stop_pos_accessor_opt<_Bounded>(__cgh, __stop_pos_storage);
+            auto __stop_pos_acc = __internal::__get_stop_pos_accessor_opt<_Bounded>(__cgh, __stop_pos_storage);
             __cgh.parallel_for<_KernelName...>(__nd_range, [=, *this](sycl::nd_item<1> __ndi)
                     [[_ONEDPL_SYCL_REQD_SUB_GROUP_SIZE_IF_SUPPORTED(__get_reduce_then_scan_req_sg_sz_device())]] {
                 const __dpl_sycl::__sub_group __sub_group = __ndi.get_sub_group();
@@ -1985,7 +1890,7 @@ struct __parallel_reduce_then_scan_reduce_submitter<_Bounded, __is_inclusive, __
                         __temp_ptr[__start_id + (__max_num_sub_groups_local - 1)] = __summary_carry.__get_cref();
                 }
 
-                if constexpr (!__is_no_stop_pos_acc_v<decltype(__stop_pos_acc)>)
+                if constexpr (!__internal::__is_no_stop_pos_acc_v<decltype(__stop_pos_acc)>)
                 {
                     if (__block_num == 0 && __ndi.get_global_linear_id() == 0)
                     {
@@ -2009,21 +1914,6 @@ struct __parallel_reduce_then_scan_reduce_submitter<_Bounded, __is_inclusive, __
     const bool __use_subgroup_ops;
 };
 
-template <typename _StopPosStorage, typename = void>
-struct __stop_pos_storage_type_selector
-{
-    using type = std::size_t;
-};
-
-template <typename _StopPosStorage>
-struct __stop_pos_storage_type_selector<_StopPosStorage, std::void_t<typename _StopPosStorage::type>>
-{
-    using type = typename _StopPosStorage::type;
-};
-
-template <typename _StopPosStorage>
-using __stop_pos_storage_type_selector_t = typename __stop_pos_storage_type_selector<_StopPosStorage>::type;
-
 // Storage for the stop position(s) of a bounded operation: the final and/or the first out-of-bounds
 // position in the source ranges. Allocated only when _Bounded=true.
 // Must be device USM because the kernel updates it with device-scope atomics (fetch_max); such
@@ -2037,158 +1927,6 @@ using __transform_reduce_then_scan_result_t =
                        std::tuple<sycl::event, __combined_storage<_ValueType>,
                                   __transform_reduce_then_scan_stop_pos_storage_t<_StopPosType>>,
                        std::tuple<sycl::event, __combined_storage<_ValueType>>>;
-
-template <bool _Bounded, typename _GenScanInput, typename _StopPosStorage>
-struct __parallel_reduce_then_scan_stop_oob_pos_tools
-{
-    using __storage_data_t = __stop_pos_storage_type_selector_t<_StopPosStorage>;
-
-    // Describes whether we have a final-position type in the storage or not
-    static constexpr bool __has_src_final_pos =
-        oneapi::dpl::__ranges::__internal::__has_final_pos_type_v<__storage_data_t>;
-
-    // Describes final position type (if any) in the storage
-    using __src_final_pos_t = oneapi::dpl::__ranges::__internal::__final_pos_type_selector_t<__storage_data_t>;
-
-    // Describes OOB position type
-    using __oob_pos_t =
-        std::conditional_t<__detect_oob_in_two_steps_v<_GenScanInput>, std::uint16_t, __src_final_pos_t>;
-
-    static __oob_pos_t
-    __create_initial_oob_pos()
-    {
-        if constexpr (std::is_arithmetic_v<__oob_pos_t>)
-            return std::numeric_limits<__oob_pos_t>::max();
-        else
-            return {};
-    }
-
-    static bool
-    __is_eq_to_initial_oob_pos(const __oob_pos_t& __pos)
-    {
-        return __pos == __create_initial_oob_pos();
-    }
-
-    template <typename _Group, typename _FinalPosType>
-    static _FinalPosType
-    __reduce_max_final_pos(_Group __group, _FinalPosType& __last_idxs_in_this_wi)
-    {
-        if constexpr (std::is_arithmetic_v<std::decay_t<_FinalPosType>>)
-        {
-            return __dpl_sycl::__reduce_over_group(__group, __last_idxs_in_this_wi,
-                                                   __dpl_sycl::__maximum<_FinalPosType>());
-        }
-        else
-        {
-            using TField0 = std::decay_t<decltype(std::get<0>(__last_idxs_in_this_wi))>;
-            using TField1 = std::decay_t<decltype(std::get<1>(__last_idxs_in_this_wi))>;
-
-            return {__dpl_sycl::__reduce_over_group(__group, std::get<0>(__last_idxs_in_this_wi),
-                                                    __dpl_sycl::__maximum<TField0>()),
-                    __dpl_sycl::__reduce_over_group(__group, std::get<1>(__last_idxs_in_this_wi),
-                                                    __dpl_sycl::__maximum<TField1>())};
-        }
-    }
-
-    template <typename __FinalAndOOBPosAcc, typename _OOBPosType>
-    static void
-    __update_oob_pos(__FinalAndOOBPosAcc& __final_and_oob_pos_acc, const _OOBPosType& __oob_pos)
-    {
-        auto& __final_and_oob_pos = __final_and_oob_pos_acc.__data()[0];
-
-        // No synchronization needed because OOB position may be reached only in a single work-item
-        if constexpr (std::is_arithmetic_v<std::decay_t<_OOBPosType>>)
-        {
-            // The __final_and_oob_pos really is simple position value
-            __final_and_oob_pos = __oob_pos;
-        }
-        else
-        {
-            // The __final_and_oob_pos really is _SetOpFinalAndOOBPosType
-            __final_and_oob_pos.__oob_pos = __oob_pos;
-        }
-    }
-
-    // Device-scope atomic over global memory. The target must live in device USM
-    // (see __transform_reduce_then_scan_stop_pos_storage_t); a device-scope atomic on host USM
-    // triggers an atomic access violation on device.
-    template <typename _FieldT>
-    using _StopPosFieldAtomicRefT =
-        sycl::atomic_ref<std::decay_t<_FieldT>, sycl::memory_order::relaxed, sycl::memory_scope::device,
-                         sycl::access::address_space::global_space>;
-
-    template <typename __FinalAndOOBPosAcc, typename _FinalPosType>
-    static void
-    __update_final_pos(__FinalAndOOBPosAcc& __final_and_oob_pos_acc, const _FinalPosType& __final_pos)
-    {
-        // The __final_and_oob_pos really is _SetOpFinalAndOOBPosType
-        auto& __final_and_oob_pos = __final_and_oob_pos_acc.__data()[0];
-
-        // We need atomic access here as multiple work-items can setup reached final position and update it concurrently.
-        if constexpr (std::is_arithmetic_v<std::decay_t<decltype(__final_and_oob_pos)>>)
-        {
-            _StopPosFieldAtomicRefT<_FinalPosType>(__final_and_oob_pos).fetch_max(__final_pos);
-        }
-        else
-        {
-            auto& __final_pos_field0 = std::get<0>(__final_and_oob_pos.__final_pos);
-            auto& __final_pos_field1 = std::get<1>(__final_and_oob_pos.__final_pos);
-
-            using _FinalPosType0 = std::decay_t<decltype(__final_pos_field0)>;
-            using _FinalPosType1 = std::decay_t<decltype(__final_pos_field1)>;
-
-            _StopPosFieldAtomicRefT<_FinalPosType0>(__final_pos_field0).fetch_max(std::get<0>(__final_pos));
-            _StopPosFieldAtomicRefT<_FinalPosType1>(__final_pos_field1).fetch_max(std::get<1>(__final_pos));
-        }
-    }
-
-    template <typename __oob_pos_t>
-    static auto
-    __create_on_oob_reached(std::size_t& __start_id_reached, std::size_t& __start_id_reached_on_oob,
-                            __oob_pos_t& __oob_detected)
-    {
-        return [&__start_id_reached, &__start_id_reached_on_oob, &__oob_detected](__oob_pos_t __id) {
-            __start_id_reached_on_oob = __start_id_reached;
-            __oob_detected = __id;
-        };
-    }
-
-    template <typename _FinalPosType>
-    static auto
-    __create_final_pos_saver(_FinalPosType& __src_final_pos)
-    {
-        if constexpr (__has_src_final_pos)
-            return [&__src_final_pos](_FinalPosType __final_pos) { __src_final_pos = __final_pos; };
-        else
-            return __no_callback_tag{};
-    }
-
-    template <typename _OOBPositionT, typename _GenScanInputArg>
-    static std::conditional_t<__detect_oob_in_two_steps_v<_GenScanInput>, __src_final_pos_t, _OOBPositionT>
-    __finalize_oob_detected(_OOBPositionT __detected_oob_pos, const std::size_t __start_id_reached_on_oob,
-                            _GenScanInputArg __gen_scan_input)
-    {
-        if constexpr (__detect_oob_in_two_steps_v<_GenScanInput>)
-        {
-            __src_pos_capturing_temp_data<__src_final_pos_t> __pos_catcher(__detected_oob_pos);
-            __gen_scan_input(__start_id_reached_on_oob, __pos_catcher, __no_callback_tag{});
-            return __pos_catcher.__get_saved_src_pos();
-        }
-        else
-        {
-            return __detected_oob_pos;
-        }
-    }
-
-    static auto
-    __create_src_final_pos_sg_container()
-    {
-        if constexpr (_Bounded && __has_src_final_pos)
-            return __src_final_pos_t{};
-        else
-            return 0;
-    }
-};
 
 template <bool _Bounded, bool __is_inclusive, bool __is_unique_pattern_v, typename _ScanOpsTag, typename _ReduceOp,
           typename _GenScanInput, typename _ScanInputTransform, typename _WriteOp, typename _InitType,
@@ -2253,7 +1991,7 @@ struct __parallel_reduce_then_scan_scan_submitter<_Bounded, __is_inclusive, __is
             auto __temp_acc = __get_accessor(sycl::read_write, __scratch_container, __cgh);
             auto __res_acc =
                 __get_result_accessor(sycl::write_only, __scratch_container, __cgh, __dpl_sycl::__no_init{});
-            auto __stop_pos_acc = __get_stop_pos_accessor_opt<_Bounded>(__cgh, __stop_pos_storage);
+            auto __stop_pos_acc = __internal::__get_stop_pos_accessor_opt<_Bounded>(__cgh, __stop_pos_storage);
 
             __cgh.parallel_for<_KernelName...>(__nd_range, [=, *this](sycl::nd_item<1> __ndi)
                     [[_ONEDPL_SYCL_REQD_SUB_GROUP_SIZE_IF_SUPPORTED(__get_reduce_then_scan_req_sg_sz_device())]] {
@@ -2458,7 +2196,7 @@ struct __parallel_reduce_then_scan_scan_submitter<_Bounded, __is_inclusive, __is
                     __group_start_id + (std::size_t{__get_sub_group_base(__ndi)} * __inputs_per_item);
                 std::size_t __start_id = __subgroup_start_id + __sub_group_local_id;
 
-                using _PosTools = __parallel_reduce_then_scan_stop_oob_pos_tools<_Bounded, _GenScanInput, _StopPosStorage>;
+                using _PosTools = __internal::__parallel_reduce_then_scan_stop_oob_pos_tools<_Bounded, _GenScanInput, _StopPosStorage>;
                 auto __src_final_pos_sg = _PosTools::__create_src_final_pos_sg_container();
 
                 if (__sub_group_id < __active_subgroups)
@@ -2512,7 +2250,8 @@ struct __parallel_reduce_then_scan_scan_submitter<_Bounded, __is_inclusive, __is
                     }
                     else
                     {
-                        __call_scan_through_elements_helper(__no_callback_tag{}, __no_callback_tag{});
+                        __call_scan_through_elements_helper(__internal::__no_callback_tag{},
+                                                            __internal::__no_callback_tag{});
                     }
                 }
 
@@ -2747,7 +2486,7 @@ __parallel_transform_reduce_then_scan_impl(sycl::queue& __q, const std::size_t _
         if constexpr (_Bounded)
             return __transform_reduce_then_scan_stop_pos_storage_t<_StopPosInitState>(__q, 1);
         else
-            return __no_stop_pos_acc_tag{};
+            return __internal::__no_stop_pos_acc_tag{};
     };
     auto __stop_pos_storage = __create_stop_pos_storage_opt(__q);
